Factor CREATE_HAS_TYPE_ALIAS into its own file
