+++ src/NumericalAlgorithms/DiscontinuousGalerkin/Protocols.hpp
@@ -6,6 +6,7 @@
 #include "DataStructures/DataBox/DataBoxTag.hpp"
 #include "Utilities/TMPL.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/CreateHasTypeAlias.hpp"
 
 namespace dg {
 /// \ref protocols related to Discontinuous Galerkin functionality
