+++ tests/Unit/Utilities/Test_ProtocolHelpers.cpp
@@ -7,6 +7,7 @@
 
 #include "Utilities/ProtocolHelpers.hpp"
 #include "Utilities/TypeTraits.hpp"
+#include "Utilities/TypeTraits/CreateHasTypeAlias.hpp"
 #include "tests/Unit/ProtocolTestHelpers.hpp"
 
 namespace {
